test(ui-info-state): use @support/jest-shared
Review feedback: @support/jest-shared landed on develop after this branch was
cut and is the config for shared/* packages, while @support/jest-features-flow
is scoped to features/flow/*. Both sibling packages already moved —
ui-qr-code to createSharedUiJestConfig, ui-queued-bottom-sheet to
createSharedUiNativeJestConfig — so this was the last shared/ui-* holdout.

createSharedUiJestConfig covers this package's web+native split directly and
ships the safe-area-context stub, so that local mock is gone.

Banner still needs a package-specific stub: the shared native stub renders its
description but not its title, and the web stub has no Banner case at all,
while InfoState sets both and its tests assert both. The two remaining mocks
now delegate to @support/jest-shared through a Proxy and override only Banner,
instead of forking the whole stub as before — so they cannot drift from the
shared implementation. Offering upstream separately; if Banner title rendering
moves into @support/jest-shared, this package drops to the same two-line config
as ui-qr-code.

Also adds @ledgerhq/test-quarantine, which the shared config's reporters and
jest-retries setup require.
