--- shared/ui-info-state/jest/mocks/safe-area-context.js
@@ -1,4 +1,0 @@
-module.exports = {
-  __esModule: true,
-  useSafeAreaInsets: () => ({ top: 0, bottom: 0, left: 0, right: 0 }),
-};
